Funky stuff. Now the folder list is filled _very_ quickly. No couple- of-second pause before you see your folders.
This means Empath now starts up incredibly quickly !

To do this, I avoid doing a recursive read immediately. This was
slowing things down because to find any subdirs of a dir, you have
to wait for QDir to give you the list. This can take absolutely ages,
as it has to stat() each file in the dir. When your Maildir has
10,000 files, that's a long time to find dirs.

So I cache the folder list, and at startup do some quick checks to
see if any of the folders do not exist any more. New folders (ones
that have been added since last run) are found in the recursive
read, which is run _after_ the cached list (minus and that disappeared)
has been shown.

EmpathIndex is now better at knowing when it is dirty.

No debug message when a Content-Type header is illegal. Needs looking
at further, but this was causing too much debug output.

svn path=/trunk/kdepim/; revision=42178