Removed a number of sorting functions that are obsolete now that qlistview does the sorting.
There is a serious problem that when operating on a selection of messages all those messages (I mean the entire message including its body) are loaded into main memory. This is not a good idea when operating on thousands of messages.

I've looked into fixing this but it's not so simple. Some functions like movemsg in kmheaders operate on a list of selected messages. And that list of selected messages is a list of whole messages not just message headers.

Even if I fix those functions I still need to know when a message is being referenced. I can audit the code and modify it as needed, but then I will still be relying on an implicit contract. Instead it would be best to modify KMFolder::getMsg so that it returns a referenced counted KMMessage wrapper by value.

svn path=/trunk/kdenetwork/kmail/; revision=47734