Work around possible source incompatible change in QT between 2.1 and 2.1.1
I suspect in older versions of qt aStr::date() returned a pointer to an empty
string when aStr is the null string but now it returns a null pointer.

Anyway suddenly KMail crashes due to various functions no longer being able
to handle null qstrings. So I want to fix this in KMail now.

Guess I should installed a old version of QT and check it out some more.

svn path=/trunk/kdenetwork/kmail/; revision=55272