Changed name of default identity from "unknown" to "Default".
I also i18n'd the code, and tried to handle the case where the user
changes language preferences (so don't save any i18n strings into the
config file). Actually for this last case I slacked off a bit when it
got to kmcomposewin but with any luck the user won't be affected.

svn path=/trunk/kdenetwork/kmail/; revision=59742