The about page.
Sorry for not waiting until the patch is reviewed. I don't know when core-devel
comes to live again and how long it takes until someone takes the time to
review it.
Since the release is waiting for that and I was asked to do it as quickly as
possible I commit it now.

svn path=/trunk/kdenetwork/kmail/; revision=81737