Ok, now it's time to commit :-)
In short: huge code changes (compared to the size of KNotes ;-) and some internal redesign.

GUI: * the notes are now saved on FocusOutEvent to reduce the possibility of data loss (Bug#19851)
     * restructured the menu
     * most important: rewrote and extended the DCOP interface. KNotes now provides methods to
       work fine with PDA sync utils, ATM that's KPilot. The number of apps that want to sync with knotes
       is not limited.

Furthermore the design should be much cleaner now and so should be the code.
There's just one tiny design issue left that makes things look less perfect. IMHO that one
could be solved by storing the notes in a xml file, I'll see...

Michael

svn path=/trunk/kdeutils/knotes/; revision=88599