"kde" send protocol, to send using kio_smtp, according to the KDE global email settings.
Coupled with some GUI configurabilty, and modification of the KMail identity
settings, this could make KMail useful for a lot more people, people who
need to use SMTP servers that KMail currently can't talk to (because of SMTP
auth or a proxy or whatnot).

In the meantime it's still configurable through the kmailrc file, so it's
an easter egg until then.

svn path=/trunk/kdenetwork/kmail/; revision=91911