Mail should now be pretty robust when feed corrupted .*.index.sorted files
The .index.sorted file is basically a dump of the current state of
kmheaders, but it can be incrementally appended.

It contains a list of item ids, and parent ids (where id is an index number
into a kmfolder).

I now sanity check item id and parent id to make sure they are in range for
the corresponding kmfolder. I also check for duplicate ids.

The file also contains keys (used for sorting in kmheaders), I now truncate
the key length to 8k when writing, and check for a sane key length when
reading. 8k should be long enough for a key, if not all that goes wrong is
sorting is wrong, and this change prevents abnormal application termination
due to trying to allocate gigs of memory.

I also notice a bug whereby realloc was being called unnecessarily often,
and fixed that.

svn path=/trunk/kdenetwork/kmail/; revision=102493