When a single imap folder is checked while a check of the same account is going on, the following happend:
- remember the complete list of folder to check
- set the list to the single folder to check
- check it
- queue check, because there is already one running
- reset folder list
- running check finishes
- queued check starts but with the full list of folders, not the single
  folder that should have been checked.

This behaviour made changing imap folders while a check is going on rather
prohibitive, as each folder selection triggers a check of that folder.

To prevent this, keep a list of folders to check once the currently running
check is done, append the folder to be checked, if its account is being
check currently, and connect to the checkFinished signal to trigger a check
of those queued folders then.

svn path=/trunk/kdepim/; revision=230380