Always print the connection lines of the todo tree. so far, only the position of the parent todo was recorded, but if the child was on a new page, this didn't work. I now added a variable that stores if the parent todo is printed on the same page or not. If a page break occurs, I draw all vertical lines of all parent todos until the page break.
Also set font to strike out for finished todos (this should be made optional). Currently, the strike out is not printed, but the items are underlines?!?!?! I set the strike out property using QFont::setStrikeOut(true), so it is actually supposed to strike them out.
So far, a line was explicitely printed to strike the items out (but that line was far too high)

svn path=/trunk/kdepim/; revision=241733