Extended the KOAlternativeLabel to use three strings of (increasing) length. If the agenda column is wide enough, the full date string is shown (e.g. Friday, September 25, 2003), otherwise the current text (Fri 25), or if even that is too long, just the day (25).
Also, add tooltips to the holiday headers of the agenda columns (but only if the whole text doesn't fit). I kuse the KOAlternativeLabel for this (using three identical strings. yes, it's a hack, I know. But the label checks if the longest text fits and in that case doesn't show a tooltip, otherwise the tooltip shows the longest string.).

svn path=/trunk/kdepim/; revision=250015