Editing an event no longer removes it from KOrganizer's list view mode.
We kept track of all list view items in a QMap and checked the existence using this map. Now if an item was edited, the old one was removed from the list view and then the new was added to the view. We just forgot to update the QMap when the item was removed, thus the view thought the item was already in the list, and the new one wasn't added.

CCMAIL: 62719-done@bugs.kde.org

svn path=/trunk/kdepim/; revision=254667