Patch by Ladislav Strojil for KMail POP filters so that the confirmation dialog allows keyboard usage.
Summary:
Added the ability to use keyboard for action selection
Cleaned the code a little bit - removed some dead code
(there's more to clean - key() method seems useless)
The old code kept track of which column corresponds to which action but it was
not done properly. IMHO this is not worth the extra work. I have changed the
code to expect the action columns to be the first three (and added an assert
that might catch future changes).
The patch is still not perfect, TAB key does not behave itself. I will have a
look at that later.

svn path=/trunk/kdepim/; revision=256551