Fix Aegypten issue #1103.
Checkboxes for sign/encrypt attachment in new message is rendered incorrectly
sometimes. The reason was that paintCell(), which positioned the widget is not
called called for items shown off-sreen. This results in the unlayouted Checkboxes
stacking up in the right top corner.  The fix is simple: hide() the items as long
as they are not needed.

As a nice bonus, I gave the checkbox background the proper color.

svn path=/trunk/kdepim/; revision=266161